Change naming and requiring conventions
After thinking about this, I think that files should be named based on
the class name. Adding _controller or _view suffixes may seem useless as
you actually don't need to rely on file names, but it makes it much
easier to quickly distinguish between 2 files without looking at the
path and it also seems like a nice practice for all of the files - you
don't need to think if you want to leave or remove suffixes in any type
of file.

Also, requiring files in one place seems easier than doing it in files
that depend on given file.