Get rid of /api/bootstrap
I've created fixtures for FeedSection and Todo models, for now one feed
section with one todo is loaded on application start. Thanks to that, I
can get rid of /api/bootstrap and adapter's code and just rely of
ember-data. Items kept by feed section are a bit 'raw' for now, but this
can be revised later.