Fix router and implement users/:user_id route
Router was unable to load a route after doing authentication step. This
happend because it was always trying to start wit loading state and then
it went to authentication again and again. Moving to some other stater
('root' in this situation) fixes the issue, although I suppose such
scenario should be implemented in router.

I also added first url in todo's view, which shows user.