Rename contact to reference in Todo
In one of the previous commits I implemented parts of todo template that
referred to isCall, I used contact as direct association on todo. In
fact, contact should stay being reference. For now it's just association
of type Radium.Contact, which makes it work only in this case - it will
be fixed while implementing more items.