Fix showing mini-loader and make feed less bouncy
When loading next or previous feed sections, it's best to wait for
rendering loaded items. That way we don't start loading new items until
view is inserted into the DOM, so there should be no situation, when a
few views are inserted at once, which can look flaky as list can also be
reordered for sorting.

I also ensured that loader is shown in integration tests.