Change a way of handling feed scrolling
Previously scrolling relied only on server response when it comes to
deciding what should be next/previus item. App was firing requests to
get data after or before given date and it was server's responsibility
to respond with right data.

This commit gives such responsibility to the client. Application keeps
track of what are the next and previous entries for each item. When user
tries to scroll up or down, application will ask for specific record.
That way we can avoid firing unneeded requests to server (because we
know what's the next/previous record up-front and we know if it's
already loaded) and we don't have to care about handling "stop" for
scrolling - feed will now that next/previous value is null and it will
not trigger any request.