Fix problem with adding todo and filtering
After adding simple version of filtering, whenever a new item was added
to collection, it was rendered twice. This was caused probably by naive
implementation - whenever collection length changed, it was regenerated,
which probably triggered observer that adds DOM elements twice.

Current implementation is done using filter mixin that was attached in
one of the pull requests for ember. I'm not really happy with that,
because I think that implementation is not the best, but it will be
enough for now and it will allow to move on to other things.