Move sorting directly to expandable array
Previously, sorting was used in controller, which usually is a nice
thing, but before making next changes, it's much better to make array
sortable. Ids will not change anyway, so we don't need to sort *after*
they're placed in correct position and we don't need to worry about
overwriting arrangedContent.