Reorganize FixtureAdapter
app/javascripts/adapters is a new folder. It will hold all our adapter
related things. We will eventually need two adapters: one for
testing/dev and the other production. This folder structure is there to
isolate parts of the application as well organize the individual parts.

Radium.FixtureAdapter has been broken up into multiple parts. There are
files for the individual query methods, maps, transforms, and finally
the serializer. This should make this code easier to understand and
modify. It should also reinforce the fact that Radium.FixtureAdapter &
Radium.FixtureSerializer are tightly coupled and should only be used in
tandem. Their semantics should not leak to other parts of the
application